StorytellingPage
I created a cube (will later be replaced by the storytelling island), when you click on the cube you get redirected to the storyTellingPage where you see the cube and the ocean that were in the Home page in the background with a transparent pop up filling the screen with text on it telling the story of The Ocean Cleanup. When you click on the escape key you go back again to the Home page. I also created a pages system and moved the content in App.js to the HomePage.